Make AI smart
The AI now checks the grid for any openings that
may allow the player to win, and places a mark in
that opening.  It still does not attempt to win
however when it has an obvious opportunity, and
the program run indefinetly when there is a draw
because the AI does not know to stop looking.