Update logic for draws and when AI wins
Added some new code to check for draws that applies
to 2 player and AI mode.  I also modified the AI code
so that it no longer runs infinetly when it draws
or wins.  The issue was that the AI didn't know when
the game was over after it won so it just continued
trying to find possible moves (even though there
were none)